Add checked versions of Player controls
These checked versions will return Ok(false) in case the requested
action is not possible, and Ok(true) if it was possible (after asking
the player to also perform the action).

Also added can_<some_action>() methods that exposes the same information
without performing actions.

Controlling can now happen on three levels:
  - Perform the action, don't care if it had an effect.
  - Perform the action, notify if not possible.
  - Check if an action is possible, then do something about that.

Note that Ok(true) is not a guarantee that the Player did the action,
just that it indicated that it should be possible and the message was
replied to. Here is one example: checked_play() could return Ok(true)
depite not doing anything if the player is already playing. can_play()
must return true if a player is playing, because returning false would
be a contradiction.