Remove optimization TODO
I can not come up with a good way to deal with this. The `ConnPath`
requires a reference to the connection, which must also be owned by the
`Player` in order to to even handling. That is very hard to declare
without upsetting the borrow checker.

The call should be fairly cheap anyway as it just clones a few small
strings.