Rework DBusError
Now it is an enum with different variants, which makes it possible to
recover. At least it is now *easier* to recover.

Previous version of the dbus crate did not implement Sync for the Error
so it was not possible to hold on to it and implement Fail at the same
time. This has been fixed now, which is why this is now possible.