Detect TrackList changes via PropertiesChanged
Some players (like VLC) do not emit the TrackListReplaced, TrackAdded,
or TrackRemoved signals. However, they might emit the PropertiesChanged
signal which we can use to detect any changes to the track list.

Sadly, with this current approach a lot of players might get double
events being processed; first the "PropertiesChanged" signal is emitted,
then the "TrackAdded" events are emitted.

This would lead to tracks showing up multiple times as the list would be
reloaded first, then the TrackAdded events are emitted.

I'm not sure how to deal with this just yet.