Fix for song changes on streams not adding event
When listening to a stream (internet radio) the track ID and url will not
change when the played song changes so no event will be added. Checking
the title and artists fixes that.

The title is checked first because most radios will set the title to
`Artist - Title` and set the station name as the artist. Artists are checked
in case that the station sets the artist metadata correctly.
There is still the rare case where 2 songs with the same title and artists
are played back to back which will not cause an event to be added.