Change PlayerStream
Instead of holding a `Vec` of futures it now holds a `VecDeque` of bus
names, a`Connection` and the currently worked on future (if any). This
allows for a much more useful `Debug` output and prevents cloning things
unless they are needed.
Also added `Clone` for `Player` and made the `Debug` formatting for
`Mpris` and `PlayerStream` skip the `Connection` fields because it's a
lot of noise and if needed you can easily just debug print the
connection directly.