Add unique name for Player and other small changes
`Player::unique_bus_name()` will return the unique bus name of the
player if a player is connected. To make this work `Player::new()` is
back to taking a `&Mpris` so that it can copy `DBusProxy` which is used
to look up the unique name.
Other changes:
`Mpris::into_stream()` renamed to `stream_players()` because it doesn't
consume Mpris.
`PlayerStream` now also holds `DBusProxy`
`MetadataIter` renamed to `MetadataIntoIter`
Some minor formatting changes