Make disabling the internal executor easier
Added `Mpris::new_no_executor()` which disables the executor and added
the `ExecutorLoop` future that ticks the executor in a loop. This way
the user doesn't need to interact with the zbus library directly.
To make this possible `Mpris` now holds the DBusProxy inside a `OnceCell`
that gets initialized the first time some method needs it. This is
needed because when you create the `Connection` without the internal
executor trying to create the Proxy will hang. All of the Executor stuff
(including some documentation) is hidden if you enable the tokio feature.

TrackID got it's Debug changed.