Rework RawMetadata
Instead of being just an alias it's now a struct that just wraps the
HashMap. This is done so that it's possible to implement From traits
making it a lot easier to covert from the HashMap received from DBus. It
implements Deref and other traits to make it act like a HashMap.