Use custom types with zbus
zbus allows to directly use custom types in the proxy macro but do that
some of the types need to implement serde because of this serde is no
longer an optional dependency. This commit adds all of the needed
changes needed to integrate the types into zbus. This includes serde
implementations, adding `From<Value>` and `Into<Value`, implementing the
`zbus::zvariant::Type` trait and `Into<MprisError>` for the zbus Error
types.
In the cases where it was impossible to integrate the type
(`MetadataValue` and `Playlist`) in a nice way the proxy methods are
just wrapped with new methods that avoid the issue.